requirements.txt
django>=3.0
psycopg2-binary>=2.9.1
django-environ
python-dotenv